pyproject.toml
[project]
name = "puzzle-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "glfw>=2.10.0",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]
